Replace ast.literal_eval with eval
The script errored on me here. I believe `ast.literal_eval` breaks on the + operator, which `eval` does not. There are more ast.literal_eval calls, they didn't break on me, but I have not used all monitor setups, so a thorough check on all of them would be in order. I am not a Python dev so please review this change carefully.